+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Compatibility:
 * Implement `Class#attached_object` method (#3039, @andrykonchin).
 * Fix `ENV#{clone,dup}` and raise `TypeError` (#3039, @andrykonchin).
 * Fix `Coverage.supported?` and raise `TypeError` if argument is not Symbol (#3039, @andrykonchin).
+* Do not autosplat a proc that accepts a single positional argument and keywords (#3039, @andrykonchin).
 
 Performance: